feat calypte_api: migrate factories from the script to their domains folders
