+++ backend/apps/calypte_api/calypte_api/deployment/factories.py
@@ -1,0 +1,11 @@
+from deployment.models import Deployment
+
+from factory import Faker
+from factory.django import DjangoModelFactory
+
+
+class DeploymentFactory(DjangoModelFactory):
+    class Meta:
+        model = Deployment
+
+    name: Faker = Faker("word")
