+++ backend/apps/calypte_api/calypte_api/deployment/management/commands/generate_test_data.py
@@ -1,115 +1,24 @@
-import datetime
 import random
 
+from deployment.factories import DeploymentFactory
 from deployment.models import Deployment
 
+from deployment_task.factories import DeploymentTaskFactory
 from deployment_task.models import DeploymentTask
+from device.factories import DeviceFactory
 from device.models import Device
+from device_group.factories import DeviceGroupFactory
 from device_group.models import DeviceGroup
 from django.core.management.base import BaseCommand
-from factory import Faker, Iterator, Sequence, post_generation
-from factory.django import DjangoModelFactory
 from faker import Faker as RawFaker
+from firmware_info.factories import FirmwareInfoFactory
 from firmware_info.models import FirmwareInfo
 from protocol.models import Protocol, ProtocolChoices
+from tag.factories import TagFactory
 from tag.models import Tag
+from task_log_record.factories import TaskLogRecordFactory
 from task_log_record.models import TaskLogRecord
 from task_status.models import TaskStatus, TaskStatusChoices
-
-
-class DeviceGroupFactory(DjangoModelFactory):
-    class Meta:
-        model = DeviceGroup
-
-    # id: Faker = Faker("uuid4")
-    name: Faker = Faker("word")
-    protocol: Iterator = Iterator(Protocol.objects.all())
-    shared_secret: Faker = Faker(
-        "password",
-        length=64,
-        special_chars=True,
-        digits=True,
-        upper_case=True,
-        lower_case=True,
-    )
-
-
-class DeviceFactory(DjangoModelFactory):
-    class Meta:
-        model = Device
-
-    name: Faker = Faker("word")
-    serial_number: Faker = Faker(
-        "password",
-        length=64,
-        special_chars=True,
-        digits=True,
-        upper_case=True,
-        lower_case=True,
-    )
-    registration_date: Faker = Faker("date_time", tzinfo=datetime.UTC)
-    last_seen: Faker = Faker("date_time")
-
-    @post_generation
-    def tags(self, create, extracted, **kwargs):
-        if not create or not extracted:
-            return
-
-        self.tags.add(*extracted)
-
-
-class FirmwareInfoFactory(DjangoModelFactory):
-    class Meta:
-        model = FirmwareInfo
-
-    version: Sequence = Sequence(lambda n: f"v0.{n}.0")
-    serial_number: Faker = Faker(
-        "password",
-        length=64,
-        special_chars=True,
-        digits=True,
-        upper_case=True,
-        lower_case=True,
-    )
-    group: Iterator = Iterator(DeviceGroup.objects.all())
-
-
-class TagFactory(DjangoModelFactory):
-    class Meta:
-        model = Tag
-
-    name: Faker = Faker("word")
-    color: Faker = Faker("color")
-    group: Iterator = Iterator(DeviceGroup.objects.all())
-
-    @post_generation
-    def devices(self, create, extracted, **kwargs):
-        if not create or not extracted:
-            return
-
-        self.devices.add(*extracted)
-
-
-class DeploymentFactory(DjangoModelFactory):
-    class Meta:
-        model = Deployment
-
-    name: Faker = Faker("word")
-
-
-class DeploymentTaskFactory(DjangoModelFactory):
-    class Meta:
-        model = DeploymentTask
-
-    state: Iterator = Iterator(TaskStatus.objects.all())
-
-
-class TaskLogRecordFactory(DjangoModelFactory):
-    class Meta:
-        model = TaskLogRecord
-
-    message: Faker = Faker("text")
-    task: Iterator = Iterator(DeploymentTask.objects.all())
 
 
 class Command(BaseCommand):
@@ -199,9 +108,9 @@ class Command(BaseCommand):
         return list(TaskLogRecord.objects.all())
 
     def delete_all_data(self) -> None:
+        TaskLogRecord.objects.all().delete()
         DeploymentTask.objects.all().delete()
         TaskStatus.objects.all().delete()
-        TaskLogRecord.objects.all().delete()
         Deployment.objects.all().delete()
         Device.objects.all().delete()
         FirmwareInfo.objects.all().delete()
@@ -212,14 +121,14 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         self.delete_all_data()
 
-        task_statuses = self.generate_task_statuses()
-        protocols = self.generate_protocols()
+        self.generate_task_statuses()
+        self.generate_protocols()
         device_groups = self.generate_device_groups(size=10)
-        firmware_info = self.generate_firmware_info(size=30)
-        tags = self.generate_tags(size=15)
-        devices = self.generate_devices(groups=device_groups, size=50)
-        deployments = self.generate_deployments(groups=device_groups, size=10)
-        deployment_tasks = self.generate_deployment_tasks(groups=device_groups, size=10)
-        task_log_records = self.generate_task_log_records(size=1000)
+        self.generate_firmware_info(size=30)
+        self.generate_tags(size=15)
+        self.generate_devices(groups=device_groups, size=50)
+        self.generate_deployments(groups=device_groups, size=10)
+        self.generate_deployment_tasks(groups=device_groups, size=10)
+        self.generate_task_log_records(size=1000)
 
         self.stdout.write(self.style.SUCCESS("Successfully generated test data."))
