+++ backend/apps/calypte_api/calypte_api/device_group/factories.py
@@ -1,0 +1,22 @@
+from device_group.models import DeviceGroup
+
+from factory import Faker, Iterator
+from factory.django import DjangoModelFactory
+from protocol.models import Protocol
+
+
+class DeviceGroupFactory(DjangoModelFactory):
+    class Meta:
+        model = DeviceGroup
+
+    # id: Faker = Faker("uuid4")
+    name: Faker = Faker("word")
+    protocol: Iterator = Iterator(Protocol.objects.all())
+    shared_secret: Faker = Faker(
+        "password",
+        length=64,
+        special_chars=True,
+        digits=True,
+        upper_case=True,
+        lower_case=True,
+    )
