+++ backend/apps/calypte_api/calypte_api/task_log_record/factories.py
@@ -1,0 +1,13 @@
+from task_log_record.models import TaskLogRecord
+
+from deployment_task.models import DeploymentTask
+from factory import Faker, Iterator
+from factory.django import DjangoModelFactory
+
+
+class TaskLogRecordFactory(DjangoModelFactory):
+    class Meta:
+        model = TaskLogRecord
+
+    message: Faker = Faker("text")
+    task: Iterator = Iterator(DeploymentTask.objects.all())
